update to clisops 0.2.0
