+++ .gitignore
@@ -46,6 +46,7 @@ coverage.xml
 *.cover
 .hypothesis/
 .pytest_cache/
+tests/_outputs
 
 # Translations
 *.mo
