+++ MANIFEST.in
@@ -1,11 +1,8 @@
-include AUTHORS.md
-include CONTRIBUTING.md
-include HISTORY.md
+graft daops
+include *.md
+global-exclude __pycache__
+global-exclude *.py[co]
+
 include LICENSE
-include README.md
-
-recursive-include tests *
-recursive-exclude * __pycache__
-recursive-exclude * *.py[co]
-
-recursive-include docs *.md conf.py Makefile make.bat *.jpg *.png *.gif
+include requirements.txt
+include requirements_dev.txt
