+++ daops/__init__.py
@@ -9,6 +9,10 @@ __copyright__ = "Copyright 2018 United Kingdom Research and Innovation"
 __license__ = "BSD"
 __version__ = "0.1.0"
 
-from .utils import *
-from .ops import *
-from .data_utils import *
+# from .utils import *
+# from .ops import (
+#     subset,
+# )
+# from .data_utils import (
+#     squeeze_dims
+# )
