+++ daops/utils/__init__.py
@@ -1,2 +1,1 @@
-from .fixer import *
-from .core import *
+from .core import is_characterised
