+++ environment.yml
@@ -6,8 +6,23 @@ channels:
 dependencies:
   - python=3.7
   - pip
+  # daops
+  # - xarray>=0.15
+  # - dask
+  # - netcdf4
+  # clisops
+  - numpy>=1.16
   - xarray>=0.15
-  - dask
-  - netcdf4
+  - pandas>=1.0.3
+  - cftime>=1.0.4
+  - netCDF4>=1.4
+  - poppler>=0.67
+  - fiona>=1.8
+  - geojson>=2.5.0
+  - shapely>=1.6
+  - geopandas>=0.7
+  - dask>=2.6.0
+  - bottleneck>=1.3.1,<1.4
+  - pyproj>=2.5
   # tests
   - pytest
