+++ setup.py
@@ -26,7 +26,7 @@ with open("README.md") as readme_file:
     _long_description = readme_file.read()
 
 
-requirements = [ ]
+requirements = [line.strip() for line in open("requirements.txt")]
 
 setup_requirements = ['pytest-runner', ]
 
@@ -53,9 +53,9 @@ setup(
         'Operating System :: POSIX :: Linux',
         'Programming Language :: Python',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
+        'Programming Language :: Python :: 3.8',
         'Topic :: Security',
         'Topic :: Internet',
         'Topic :: Scientific/Engineering',
@@ -69,24 +69,18 @@ setup(
 
     # This qualifier can be used to selectively exclude Python versions -
     # in this case early Python 2 and 3 releases
-    python_requires='>=3.5.0',
-    entry_points={
-        'console_scripts': [
-            'daops=daops.cli:main',
-        ],
-    },
+    python_requires='>=3.6.0',
     install_requires=requirements,
     long_description=_long_description,
     long_description_content_type='text/markdown',
-
     include_package_data=True,
     keywords='daops',
     name='daops',
-    packages=find_packages(include=['daops', 'daops.*']),
+    packages=find_packages(),
     setup_requires=setup_requirements,
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/ellesmith88/daops',
     version=_package_version,
-    zip_safe=False,
+    # zip_safe=False,
 )
