+++ tests/test_operations/test_subset.py
@@ -2,7 +2,7 @@ import pytest
 
 from daops import ops
 
-from tests._common import CMIP5_ARCHIVE_BASE
+from tests._common import CMIP5_ARCHIVE_BASE, TESTS_OUTPUTS
 
 CMIP5_IDS = [
     'cmip5.output1.INM.inmcm4.rcp45.mon.ocean.Omon.r1i1p1.latest.zostoga',
@@ -17,35 +17,35 @@ def setup_module(module):
 
 def test_subset_zostoga_with_fix():
     result = ops.subset(CMIP5_IDS[0],
-                          time=('2085-01-01', '2120-12-30'),
-                          data_root_dir=CMIP5_ARCHIVE_BASE,
-                          output_dir='outputs')
-    assert result.file_paths == ['outputs/output.nc']
+                        time=('2085-01-01', '2120-12-30'),
+                        data_root_dir=CMIP5_ARCHIVE_BASE,
+                        output_dir=TESTS_OUTPUTS)
+    assert 'output.nc' in result.file_paths[0]
 
 
 def test_subset_t():
     result = ops.subset(CMIP5_IDS[1],
-                          time=('2085-01-01', '2120-12-30'),
-                          data_root_dir=CMIP5_ARCHIVE_BASE,
-                          output_dir='outputs')
-    assert result.file_paths == ['outputs/output.nc']
+                        time=('2085-01-01', '2120-12-30'),
+                        data_root_dir=CMIP5_ARCHIVE_BASE,
+                        output_dir=TESTS_OUTPUTS)
+    assert 'output.nc' in result.file_paths[0]
 
 
 @pytest.mark.skip('FAILS with TypeError. Needs fixing!')
 def test_subset_t_y_x():
     result = ops.subset(CMIP5_IDS[1],
-                          time=('2085-01-01', '2120-12-30'),
-                          space=('-20', '-10', '5', '15'),
-                          data_root_dir=CMIP5_ARCHIVE_BASE,
-                          output_dir='outputs')
+                        time=('2085-01-01', '2120-12-30'),
+                        space=('-20', '-10', '5', '15'),
+                        data_root_dir=CMIP5_ARCHIVE_BASE,
+                        output_dir=TESTS_OUTPUTS)
 
 
 @pytest.mark.skip('FAILS - needs fixing by bringing range into calendar range')
 def test_subset_t_with_invalid_date():
     result = ops.subset(CMIP5_IDS[1],
-                          time=('2085-01-01', '2120-12-31'),
-                          data_root_dir=CMIP5_ARCHIVE_BASE,
-                          output_dir='outputs')
+                        time=('2085-01-01', '2120-12-31'),
+                        data_root_dir=CMIP5_ARCHIVE_BASE,
+                        output_dir=TESTS_OUTPUTS)
 
 
 def teardown_module(module):
