+++ tests/test_xarray/test_xarray_aggregation.py
@@ -21,6 +21,8 @@ import pytest
 import tempfile
 import pathlib
 
+from .._common import TESTS_OUTPUTS
+
 test_files = [
     "tests/mini-esgf-data/test_data/badc/cmip5/data/cmip5/output1/MOHC/HadGEM2-ES/rcp85/mon/atmos/Amon/r1i1p1/latest/tas/tas_Amon_HadGEM2-ES_rcp85_r1i1p1_200512-203011.nc",
     "tests/mini-esgf-data/test_data/badc/cmip5/data/cmip5/output1/MOHC/HadGEM2-ES/rcp85/mon/atmos/Amon/r1i1p1/latest/tas/tas_Amon_HadGEM2-ES_rcp85_r1i1p1_203012-205511.nc",
@@ -32,7 +34,7 @@ F1, F2, F3 = test_files
 
 # Functions to make modified NC files
 # need to make files temporary files
-def _make_nc_modify_var_attr(nc_path, var_id, attr, value, path='tests/test_outputs/'):
+def _make_nc_modify_var_attr(nc_path, var_id, attr, value, path=TESTS_OUTPUTS):
     ds = _open(nc_path)
     ds[var_id].attrs[attr] = value
     ds.to_netcdf(os.path.join(path, 'tas_modify_var_attr.nc'))
@@ -40,7 +42,7 @@ def _make_nc_modify_var_attr(nc_path, var_id, attr, value, path='tests/test_outp
     return tmp_path
 
 
-def _make_nc_modify_global_attr(nc_path, attr, value, path='tests/test_outputs/'):
+def _make_nc_modify_global_attr(nc_path, attr, value, path=TESTS_OUTPUTS):
     ds = _open(nc_path)
     ds.attrs[attr] = value
     # ds.to_netcdf(path=tmp_path.mkdir("test_dir").join("modify_var_attr.nc"))
@@ -49,7 +51,7 @@ def _make_nc_modify_global_attr(nc_path, attr, value, path='tests/test_outputs/'
     return tmp_path
 
 
-def _make_nc_modify_var_id(nc_path, old_var_id, new_var_id, path='tests/test_outputs/'):
+def _make_nc_modify_var_id(nc_path, old_var_id, new_var_id, path=TESTS_OUTPUTS):
     ds = _open(nc_path)
     ds = ds.rename({old_var_id: new_var_id})
     ds.to_netcdf(path=os.path.join(path, 'tas_modify_var_id.nc'))
@@ -57,7 +59,7 @@ def _make_nc_modify_var_id(nc_path, old_var_id, new_var_id, path='tests/test_out
     return tmp_path
 
 
-def _make_nc_modify_fill_value(nc_path, var_id, fill_value, path='tests/test_outputs/'):
+def _make_nc_modify_fill_value(nc_path, var_id, fill_value, path=TESTS_OUTPUTS):
     ds = _open(nc_path)
     ds[var_id].encoding["_FillValue"] = fill_value
     ds.tas.encoding["missing_value"] = fill_value
